Abstracted readers/writers for measurements files
There's an issue with how measurements files were recorded that caused files
written on 64-bit machines to be unreadable on 32-bit machines.  Rethinking the
file format makes me want to change it altogether.  The new format will be more
like a dump of a single table so it should be simpler to read into matlab,
etc...  For backwards compatibility, I'm using the abstraction pattern I used
for the whiskers files.

  - builds
  - basic tests passed (traj.py) for old format
  - v0 is the old format
  - v1 (measV1) is the new default format
    The new format isn't really any different than v0, although there is a
    header specifying the format.

git-svn-id: https://subversion.int.janelia.org/myerslab/projects/whisker/branches/2009-05-25@513 fbdb9df1-db47-0410-955a-fa45b24b025c